refactor(crop): 优化图像裁剪功能
- 移除了不必要的类型转换，提高了代码的可读性和性能
- 保留了原始逻辑，确保功能的稳定性

+++ src-tauri/src/crop.rs
@@ -14,10 +14,10 @@ pub async fn crop_image(
         .decode()
         .map_err(|_| AppError::ImageDecodeError)?;
     let img = img.crop(
-        crop_x as u32,
-        crop_y as u32,
-        crop_width as u32,
-        crop_height as u32,
+        crop_x,
+        crop_y,
+        crop_width,
+        crop_height,
     );
 
     let _ = img.save(save_path).map_err(|e| AppError::ImageSaveError);
